+++ presentation/src/main/java/com/example/dms_android/feature/register/screen/id/SetIdScreen.kt
@@ -1,14 +1,18 @@
-package com.example.dms_android.feature.register.screen
+package com.example.dms_android.feature.register.screen.id
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -16,6 +20,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
@@ -30,12 +35,15 @@ import com.example.dms_android.R
 
 @Preview
 @Composable
-fun EnterEmailScreen() {
-    var value by remember { mutableStateOf(String()) }
+fun SetIdScreen() {
+    var gradeValue by remember { mutableStateOf("") }
+    var classRoomValue by remember { mutableStateOf("") }
+    var numberValue by remember { mutableStateOf("") }
+    var idValue by remember { mutableStateOf("") }
     Box(
         modifier = Modifier
             .fillMaxSize()
-            .background(DormColor.Gray100)
+            .background(Color.White)
     ) {
         Column(
             modifier = Modifier
@@ -46,35 +54,59 @@ fun EnterEmailScreen() {
             Image(
                 painter = painterResource(id = DormIcon.Applicate.drawableId),
                 contentDescription = null,
-                modifier = Modifier
-                    .size(49.dp),
+                modifier = Modifier.size(49.dp),
             )
             Spacer(
                 modifier = Modifier
                     .height(7.dp)
             )
-            Body4(
-                text = stringResource(R.string.enter_email_address),
-                color = DormColor.Gray600,
-            )
+            Body4(text = stringResource(R.string.set_id), color = DormColor.Gray600)
             Spacer(
                 modifier = Modifier
                     .height(60.dp)
             )
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween,
+            ) {
+                //TODO: Toast완성되면 에러처리함
+                DormTextField(
+                    value = gradeValue,
+                    onValueChange = { gradeValue = it },
+                    hint = stringResource(R.string.grade),
+                    modifier = Modifier.width(110.dp),
+                )
+                DormTextField(
+                    value = classRoomValue,
+                    onValueChange = { classRoomValue = it },
+                    hint = stringResource(R.string.class_room),
+                    modifier = Modifier.width(110.dp),
+                )
+                DormTextField(
+                    value = numberValue,
+                    onValueChange = { numberValue = it },
+                    hint = stringResource(R.string.number),
+                    modifier = Modifier.width(110.dp),
+                )
+            }
+            Spacer(
+                modifier = Modifier
+                    .height(24.dp)
+            )
             DormTextField(
-                value = value,
-                onValueChange = { value = it },
-                error = stringResource(R.string.not_valid_email_format),
-                hint = stringResource(R.string.enter_email_address),
+                value = idValue,
+                onValueChange = { idValue = it },
+                hint = stringResource(R.string.enter_id),
+                error = stringResource(R.string.using_id),
             )
             Box(
                 contentAlignment = Alignment.BottomCenter,
                 modifier = Modifier
                     .fillMaxSize()
-                    .padding(bottom = 60.dp),
+                    .padding(bottom = 60.dp)
             ) {
                 DormContainedLargeButton(
-                    text = stringResource(R.string.send_verification_code),
+                    text = stringResource(R.string.next),
                     color = DormButtonColor.Blue,
                     enabled = false,
                 ) {
